requirements.txt
certifi==2019.3.9
chardet==3.0.4
idna==2.5
python-dotenv==0.10.1
requests2==2.16.0
urllib3==1.21.1
